requirements.txt
aiogram==3.17.0
langchain==0.3.14
langchain-community==0.3.14
openai==1.59.3
pandas==2.2.3
python-dotenv==1.0.1